+++ MoreShipUpgrades/UpgradeComponents/Commands/Interns.cs
@@ -16,6 +16,8 @@ namespace MoreShipUpgrades.UpgradeComponents.Commands
 
         List<PlayerControllerB> recentlyInterned;
         internal string[] internNames, internInterests;
+        internal int currentUsages;
+        internal float revivalTimer;
         public enum TeleportRestriction
         {
             None,
@@ -29,6 +31,14 @@ namespace MoreShipUpgrades.UpgradeComponents.Commands
             recentlyInterned = new();
             internNames = AssetBundleHandler.GetInfoFromJSON("InternNames").Split(",");
             internInterests = AssetBundleHandler.GetInfoFromJSON("InternInterests").Split(",");
+            currentUsages = 0;
+            revivalTimer = 0f;
+        }
+
+        void Update()
+        {
+            if (revivalTimer > 0f)
+                revivalTimer -= Time.deltaTime;
         }
 
         [ServerRpc(RequireOwnership = false)]
@@ -55,6 +65,8 @@ namespace MoreShipUpgrades.UpgradeComponents.Commands
         [ClientRpc]
         private void ReviveTargetedPlayerClientRpc()
         {
+            currentUsages++;
+            revivalTimer = UpgradeBus.Instance.PluginConfiguration.INTERNS_INTERVAL_BETWEEN_REVIVES;
             PlayerControllerB player = StartOfRound.Instance.mapScreen.targetedPlayer;
             int health = 100;
             if (UpgradeBus.Instance.PluginConfiguration.StimpackConfiguration.Enabled)